# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 芽纪 API
  version: 0.2.0
  description: |
    MVP OpenAPI 契约，对应执行规格第 8.2 节。
    2026-08-09 原型确认后对齐 v2.0：存量导入、就诊录音、阶段小结、「此刻」与知识库。
servers:
  - url: /api/v1
paths:
...
  /families/{familyId}:
    get:
      summary: 家庭详情
      parameters:
        - $ref: "#/components/parameters/familyId"
      responses:
        "200":
          description: 家庭
    patch:
      summary: 更新家庭
      parameters:
        - $ref: "#/components/parameters/familyId"
      responses:
        "200":
          description: 家庭
...
components:
  parameters:
    familyId:
      name: familyId
      in: path
      required: true
      schema: { type: string, format: uuid }
...